requirements.txt
streamlit
torch
torchvision
Pillow
faiss-cpu
numpy
pandas